Prevent RemoteInputView from retaining NotificationEntry
It has been observed in heap dumps that the InputMethodManager can hold
a NotificationEntry in memory after the notification has been dismissed.
To prevent this, this change breaks the reference chain between the
RemoteEditText and the RemoteInputView. The cut is performed when the
RemoteInputView is detached from the window.

Fixes: 159201509
Test: manual - Click on smart reply. Take a heap dump. Verify that
NotificationEntry isn't retained in memory via InputMethodManager.

Change-Id: Idfdfa49dbc32bb2f256bc5da74dbd9eeeb7f2153